Rename anthropic.py to anthropic_api.py
